requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
vaderSentiment==3.3.2
whitenoise==5.2.0
